+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- --save-state option for commands like gg can commit\n\nThis is needed to make GitHub pipelines work
+
 ## [1.0.14] - 2024-04-05
 
 ### Added
@@ -15,5 +21,6 @@
 
 - Initial version
 
+[Unreleased]: https://github.com/inlavigo/gg/compare/1.0.14...HEAD
 [1.0.14]: https://github.com/inlavigo/gg/compare/1.0.12...1.0.14
 [1.0.12]: https://github.com/inlavigo/gg/releases/tag/1.0.12